+++ apps/client/src/environments/patch-notes.ts
@@ -1,7 +1,16 @@
 export const patchNotes = `### Bug Fixes
 
-* **ad:** disabled ad banner on desktop entirely.
-* **desktop:** fixed retainer tracking and item HQ flags.
-* **search:** fixed desktop app download button always showing.
-* **user-menu:** fixed padding for user submenu in topbar.
-* **ux:** proper hover icon on mouse pointer for back button.`;
+* **desktop:** fixed retainer inventories being reset when getting gil from them.
+* **desktop:** fixed string conversion in CN packet capture.
+* **gearset:** fixed gearset import from packet capture.
+* **layout-editor:** fixed a bug with tiers display not shown properly in preview.
+* **list:** layouts now only consider level if it's > 0.
+* **log-tracker:** fixed an issue with page selector not being displayed properly.
+* **overlay:** fixed missing scrollbar in overlays.
+* **ux:** fix empty tooltip for alarm button.
+
+
+### Features
+
+* **desktop:** new setting to hide back button in the sidebar.
+* **layout:** new filter: IS_VOYAGE.`;